[OPP-1379] trekke spring STS konfig ut i egen fil
